backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.24.1
pyspellchecker==0.7.2
textblob==0.17.1
nltk==3.8.1
SpeechRecognition==3.10.0
PyAudio==0.2.13
requests==2.31.0 